Modification de la liste d'articles et implementation des requirement rabbitMQ et token de sécurité

+++ src/main/java/com/mspr4/ordersapi/controller/OrdersController.java
@@ -1,5 +1,6 @@
 package com.mspr4.ordersapi.controller;
 
+import com.mspr4.ordersapi.messaging.OrdersEventPublisher;
 import com.mspr4.ordersapi.model.Orders;
 import com.mspr4.ordersapi.service.OrdersService;
 import org.springframework.http.ResponseEntity;
@@ -8,15 +9,16 @@ import jakarta.validation.Valid;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
-
 @RestController
 @RequestMapping("/api/orders")
 public class OrdersController {
 
     private final OrdersService service;
+    private final OrdersEventPublisher eventPublisher;
 
-    public OrdersController(OrdersService service) {
+    public OrdersController(OrdersService service, OrdersEventPublisher eventPublisher) {
         this.service = service;
+        this.eventPublisher = eventPublisher;
     }
 
     @GetMapping
@@ -34,6 +36,7 @@ public class OrdersController {
     @PostMapping
     public ResponseEntity<Orders> create(@Valid @RequestBody Orders order) {
         Orders saved = service.saveOrder(order);
+        eventPublisher.publishOrderCreated(saved); // publication de l'événement
         return ResponseEntity.status(201).body(saved);
     }
 
@@ -45,19 +48,23 @@ public class OrdersController {
                     existing.setCustomerId(order.getCustomerId());
                     existing.setOrderDate(order.getOrderDate());
                     existing.setStatus(order.getStatus());
-                    existing.setTotalAmount(order.getTotalAmount());
-                    existing.setDeliveryAddress(order.getDeliveryAddress());
                     existing.setOrderItems(order.getOrderItems());
-                    return ResponseEntity.ok(service.saveOrder(existing));
+                    Orders updated = service.saveOrder(existing);
+                    eventPublisher.publishOrderUpdated(updated); // publication
+                    return ResponseEntity.ok(updated);
                 })
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable UUID id) {
-        try {
+        var existing = service.getOrderById(id);
+        if (existing.isPresent()) {
             service.deleteOrder(id);
-        } catch (NoSuchElementException ignored) {}
-        return ResponseEntity.noContent().build(); // 204
+            eventPublisher.publishOrderDeleted(existing.get());
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
